% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{num_int}
\alias{num_int}
\alias{NumInt}
\title{Monte Carlo integration
This function performes numerical integration using monte carlo integration}
\usage{
num_int(fn, ..., a, b, N = 100, ncores = 1, cl = NULL)
}
\arguments{
\item{fn}{is a function}

\item{...}{are further arguments to be passed to fn}

\item{a}{is the lower bound}

\item{b}{is the upper bound}

\item{N}{is the sample size}

\item{ncores}{is the number of processors to call}

\item{cl}{is an object of class cluster}
}
\value{
An object of class "numint"
}
\description{
Monte Carlo integration
This function performes numerical integration using monte carlo integration
}
\references{
Boyle, E. A., Li, Y. I., & Pritchard, J. K. (2017). An Expanded View of Complex Traits: From Polygenic to Omnigenic. Cell, 169(7), 1177???1186. http://doi.org/10.1016/j.cell.2017.05.038
}
