% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint-pkg.R
\docType{package}
\name{numint-pkg}
\alias{numint-pkg}
\alias{numint-pkg-package}
\title{numint}
\description{
We add stuff up... You can access to the project
website at \url{https://github.com/USCbiostats/software-dev/tree/master/rpkgs/funnypkg}
}
\details{
A not so funny collection of functions
}
\author{
George G. Vega Yon
}
