% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{plot.numint}
\alias{plot.numint}
\title{Plotting method for num_int}
\usage{
\method{plot}{numint}(x, y = NULL, main = "Monte Carlo Integration",
  col = blues9[4], ...)
}
\arguments{
\item{x}{the x-coordinate}

\item{y}{the y-coordinate (ignored, but need tbecause the standard plot function has them)}

\item{main}{the title}

\item{col}{the color to use}

\item{...}{additional parameters to pass}
}
\value{
it returns the coordinates
}
\description{
Plotting method for num_int
}
