% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.R
\name{plot.funnypkg_addnums}
\alias{plot.funnypkg_addnums}
\title{Addnums}
\usage{
\method{plot}{funnypkg_addnums}(x, y = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{funnypkg_addnums}.}

\item{y}{Ignored.}

\item{...}{Further arguments passed to
\code{\link[graphics:plot.window]{plot.window}}.}
}
\description{
Addnums
}
