% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numint.r
\name{num_int}
\alias{num_int}
\alias{NumInt}
\alias{numint}
\alias{plot.numint}
\alias{print.numint}
\title{Monte Carlo Integration}
\usage{
num_int(fn, ..., a, b, N = 100, ncores = 1, cl = NULL)

numint(fn, ..., a, b, N = 100, ncores = 1, cl = NULL)

\method{plot}{numint}(x, y = NULL, main = "Monte Carlo Integration",
  col = blues9[4], ...)

\method{print}{numint}(x, ...)
}
\arguments{
\item{fn}{A function}

\item{...}{Further arguments to be passed to -fn-}

\item{a}{Numeric vector. Lower bound}

\item{b}{Numeric vector. Upper bound}

\item{N}{Integer scalar. Sample size}

\item{ncores}{Integer scalar. Number of cores (parallel)}

\item{cl}{An object of class cluster}

\item{x}{An object of class numint}

\item{y}{Ignored}

\item{main}{Title}

\item{col}{Color}
}
\value{
An object of class
}
\description{
This function performes numerical integration using monte carlo integration
}
\details{
\code{code}, \emph{emph}, \pkg{igraph}, \CRANpkg{igraph}
\eqn{\pi}{pi}

\deqn{}{}
}
\examples{
# Simple example -------
ans <- num_int(dnorm, mean = .5, a = -6, b = 7, N = 1e3)

# Multiple proc -----
\dontrun{

ans <- num_int(dnorm, mean = .5, a = -6, b = 7, N = 1e8,
ncores = 20)

}
}
\references{
Weisstein, Eric W. "Monte Carlo Integration." From MathWorld--A Wolfram Web Resource. \url{http://mathworld.wolfram.com/MonteCarloIntegration.html}
}
\author{
George
}
